PBI_XMLA_Chatbot_Package/Chatbot/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
azure-identity
